refactors Moved components to their own folders.
