+++ App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Alert, Button, View, Text, Image, StyleSheet, FlatList } from "react-native";
-import Header from "./components/Header";
-import ListItem from "./components/ListItem";
-import AddItem from "./components/AddItem";
+import Header from "./components/Header/Header";
+import ListItem from "./components/ListItem/ListItem";
+import AddItem from "./components/AddItem/AddItem";
 
 export default class App extends React.Component {
     state = {
